+++ pkg/vault/client/github/const.go
@@ -1,0 +1,3 @@
+package github
+
+const githubDefaultAuthPath = "auth/github/login"